Rfc2045CharPredicate moved to CharPredicates.rfc2045(), others inlined.
- HttpCharPredicates deleted.
