--- src/test/java/walkingkooka/net/http/HttpCharPredicatesTest.java
@@ -1,36 +1,0 @@
-/*
- * Copyright 2018 Miroslav Pokorny (github.com/mP1)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- *
- */
-
-package walkingkooka.net.http;
-
-import walkingkooka.test.PublicStaticHelperTestCase;
-
-import java.lang.reflect.Method;
-
-final public class HttpCharPredicatesTest extends PublicStaticHelperTestCase<HttpCharPredicates> {
-
-    @Override
-    protected Class<HttpCharPredicates> type() {
-        return HttpCharPredicates.class;
-    }
-
-    @Override
-    protected boolean canHavePublicTypes(final Method method) {
-        return false;
-    }
-}
